# Memory optimizations for resampling AR parameters and computing AR likelihoods