Features:

- Implemented computation of AR log-likelihoods using convolutions so to save memory and take advantage of GPU

Testing speed on a GPU, I see a ~400ms speedup per sampling iteration for a 15-session, ~35k frame/session dataset using convolution (8.6s / iteration) over the original method (9.08s / iteration) for computing AR log-likelihoods.

- Refactored `resample_ar_params` to run `jax.lax.map` over states rather than `vmap`'ing over states to reduce memory consumption

Testing speed on a GPU, I see a ~300ms _slowdown_ per sampling iteration for a 15-session, ~35k frame/session dataset using `jax.lax.map` (9.08s / iteration) over the original method (8.72s / iteration)

Here are some stats about memory usage in these two modes running on the CPU:
| fun type             | max mem | n_sessions |
| -------------------- | ------- | ---------- |
| vmap (cpu)     | 13.7 Gb   | 15         |
| lax.map (cpu) | 2.5 Gb   | 15         |

- Added an example python script to the `examples/` folder to show how `jax-moseq` can be run outside of a jupyter notebook. Please add comments and suggestions if you have any!